+++ packages/ee/ui-platform/src/index.ts
@@ -1,3 +1,2 @@
-export * from './lib/components/pieces-table/pieces-table.component';
 export * from './lib/ui-ee-platform.module';
 export * from './lib/lib.routes';
