+++ packages/ee/ui-platform/src/lib/components/dialogs/edit-add-piece-oauth-2-credentials-dialog/edit-add-piece-oauth-2-credentials-dialog.component.html
