--- packages/ee/ui-platform/src/lib/components/platform/platform.component.ts
@@ -1,29 +1,0 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  OnDestroy,
-  OnInit,
-} from '@angular/core';
-import { DashboardService } from '@activepieces/ui/common';
-import { Platform } from '@activepieces/ee-shared';
-import { ActivatedRoute } from '@angular/router';
-
-@Component({
-  selector: 'app-platform',
-  templateUrl: './platform.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class PlatformComponent implements OnInit, OnDestroy {
-  constructor(
-    private dashboardService: DashboardService,
-    private route: ActivatedRoute
-  ) {}
-  platform!: Platform;
-  ngOnInit() {
-    this.platform = this.route.snapshot.data['platform'];
-    this.dashboardService.enteredPlatformModule();
-  }
-  ngOnDestroy() {
-    this.dashboardService.leftPlatformModule();
-  }
-}
