+++ packages/ee/ui-platform/src/lib/lib.routes.ts
@@ -1,15 +1,15 @@
 import { Route } from '@angular/router';
-import { PlatformComponent } from './components/platform/platform.component';
+import { PlatformDashboardContainerComponent } from './pages/platform-dashboard-container/platform-dashboard-container.component';
 import { platformResolver } from './platform.resolver';
-import { ProjectsTableComponent } from './components/projects-table/projects-table.component';
-import { PlatformAppearanceComponent } from './components/platform-appearance/platform-appearance.component';
-import { PlatformSettingsComponent } from './components/platform-settings/platform-settings.component';
-import { PiecesTableComponent } from './components/pieces-table/pieces-table.component';
+import { ProjectsTableComponent } from './pages/projects-table/projects-table.component';
+import { PlatformAppearanceComponent } from './pages/platform-appearance/platform-appearance.component';
+import { PlatformSettingsComponent } from './pages/platform-settings/platform-settings.component';
+import { PiecesTableComponent } from './pages/pieces-table/pieces-table.component';
 
 export const uiEePlatformRoutes: Route[] = [
   {
     path: '',
-    component: PlatformComponent,
+    component: PlatformDashboardContainerComponent,
     children: [
       {
         path: '',
