+++ packages/ee/ui-platform/src/lib/pages/pieces-table/pieces-table.component.html
