+++ packages/ee/ui-platform/src/lib/pages/platform-appearance/platform-appearance.component.html
