+++ packages/ee/ui-platform/src/lib/pages/platform-dashboard-container/platform-dashboard-container.component.html
@@ -3,5 +3,4 @@
 
 <div class="ap-px-7.5 ap-py-5 ap-h-[calc(100vh-11.5625rem)]">
     <router-outlet></router-outlet>
-
 </div>
