+++ packages/ee/ui-platform/src/lib/pages/platform-settings/platform-settings.component.ts
