+++ packages/ee/ui-platform/src/lib/pages/platform-settings/signing-keys-table/signing-keys-table.component.html
