+++ packages/ee/ui-platform/src/lib/pages/projects-table/create-project-dialog/create-project-dialog.component.ts
