+++ packages/ee/ui-platform/src/lib/pages/projects-table/projects-table.component.ts
