+++ packages/ee/ui-platform/src/lib/pages/projects-table/update-project-dialog/update-project-dialog.component.html
